pyproject.toml
[tool.poetry]
name = "mtg-glue"
version = "0.1.0"
description = "A bunch of scripts I use to manage my MTG collection."
authors = ["Kevin McCartney"]
readme = "README.md"
packages = [
  { include = "cli", from = "projects" },
  { include = "lib", from = "projects" },
  { include = "models", from = "projects" },
  { include = "etl", from = "projects" },
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pydantic = "^2.12.5"
pyyaml = "^6.0.2"
playwright = "^1.58"
python-dotenv = "^1.0"
boto3 = "^1.35"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pylint = "^4.0.4"
black = "^26.1.0"
isort = "^7.0.0"
mypy = "^1.19.1"
flake8 = "^7.3.0"
types-pyyaml = "^6.0.12.20250915"
types-boto3 = "^1.42.82"
debugpy = "^1.8.20"

[tool.poetry.scripts]
moxfield-import = "cli.echo_mtg_to_moxfield:main"
echo-moxfield-etl = "etl.echo_moxfield_etl:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

